FinalProj/requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.3
google-auth==1.7.0
requests==2.21.0
google-auth-oauthlib==0.4.1
oauthlib==3.1.0
requests-oauthlib==1.2.0
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
